+++ src/channel/enums/i18n.ts
@@ -1,0 +1,4 @@
+export enum I18n {
+	E_NOT_UNIQUE = 'api.channel.error.not_unique',
+	E_DB_NOT_READY = 'api.channel.error.db.not_ready',
+}