+++ src/channel/enums/i18nPrefix.ts
@@ -1,3 +1,0 @@
-export enum I18nPrefix {
-	ERROR = 'API.CHANNEL.ERROR',
-}