+++ src/core/enums/HttpStatusCodes.ts
@@ -1,4 +1,0 @@
-export enum HttpStatus {
-	OK = 200,
-	BAD_REQUEST = 400,
-}