+++ src/user-account/enums/i18n.ts
@@ -1,0 +1,6 @@
+export enum I18n {
+	E_ = 'api.user-account.error.',
+	E_DB_NOT_READY = 'api.user-account.error.db.not_ready',
+	E_VALIDATION_FAILED = 'api.user-account.error.validation.failed',
+	E_EMAIL_NOT_FOUND = 'api.user-account.error.email.not_found',
+}